deleted old unsupported python implementations. removed unused files. Converted old circle examples to geometry shapes.
