--- docs/line.md
@@ -1,26 +1,0 @@
-# geometry.Line
-
-## copy()
-
-Returns a copy of itself
-
-## raycast()
-
-## collideline()
-
-Tests if two lines overlap
-
-## collidepoint()
-
-Tests if a point overlaps a line
-
-## collidecircle()
-
-Tests if a circle overlaps a line
-
-## as_frect()
-
-Returns a bounding box of itself as an FRect
-
-## update()
-
